chore: CE EE split for bottom bar UI
