+++ app/client/src/ce/components/BottomBar/ManualUpgrades.tsx
