+++ app/client/src/ce/components/BottomBar/index.tsx
