+++ app/client/src/ee/components/BottomBar/index.tsx
@@ -1,0 +1,3 @@
+import { default as CE_BottomBar } from "ce/components/BottomBar";
+
+export default CE_BottomBar;
